refactor: move functions submit_password and backend_submit_password from core to auth module
this change requires gui module to be imported in auth.py, to avoid circular imports and since gui is imported only for type checking, __future__ and TYPE_CHECKING are added to imports
user deletion is added to backend_submit_password before changing/adding the selected user
update button callbacks accordingly
