+++ dlunch/core.py
@@ -22,10 +22,8 @@ from io import BytesIO
 from PIL import Image
 from pytesseract import pytesseract
 import random
-import re
 from sqlalchemy import func, select, delete, update
 from sqlalchemy.sql.expression import true as sql_true
-from time import sleep
 
 # Graphic interface imports (after class definition)
 from . import models
@@ -1313,137 +1311,3 @@ def download_dataframe(
         )
 
     return bytes_io
-
-
-def submit_password(gi: gui.GraphicInterface, config: DictConfig) -> bool:
-    """Same as backend_submit_password with an additional check on old
-    password.
-
-    Args:
-        config (DictConfig): Hydra configuration dictionary.
-        gi (gui.GraphicInterface): graphic interface object (used to interact with Panel widgets).
-
-    Returns:
-        bool: true if successful, false otherwise.
-    """
-    # Get user's password hash
-    password_hash = auth.get_hash_from_user(pn_user(config), config=config)
-    # Get username, updated updated at each key press
-    old_password_key_press = gi.password_widget._widgets[
-        "old_password"
-    ].value_input
-    # Check if old password is correct
-    if password_hash == old_password_key_press:
-        # Check if new password match repeat password
-        return backend_submit_password(
-            gi=gi, config=config, user=pn_user(config), logout_on_success=True
-        )
-    else:
-        pn.state.notifications.error(
-            "Incorrect old password!",
-            duration=config.panel.notifications.duration,
-        )
-
-    return False
-
-
-def backend_submit_password(
-    gi: gui.GraphicInterface | gui.BackendInterface,
-    config: DictConfig,
-    user: str = None,
-    is_guest: bool | None = None,
-    is_admin: bool | None = None,
-    logout_on_success: bool = False,
-) -> bool:
-    """Submit password to database from backend but used also from frontend as
-    part of `submit_password` function.
-
-    When used for backend `is_guest` and `is_admin` are selected from a widget.
-    When called from frontend they are `None` and the function read them from
-    database using the user input.
-
-    Args:
-        gi (gui.GraphicInterface | gui.BackendInterface): graphic interface object (used to interact with Panel widgets).
-        config (DictConfig): Hydra configuration dictionary.
-        user (str, optional): username. Defaults to None.
-        is_guest (bool | None, optional): guest flag (true if guest). Defaults to None.
-        is_admin (bool | None, optional): admin flag (true if admin). Defaults to None.
-        logout_on_success (bool, optional): set to true to force logout once the new password is set. Defaults to False.
-
-    Returns:
-        bool: true if successful, false otherwise.
-    """
-    # Check if user is passed, otherwise check if backend widget
-    # (password_widget.object.user) is available
-    if not user:
-        username = gi.password_widget._widgets["user"].value_input
-    else:
-        username = user
-    # Get all passwords, updated at each key press
-    new_password_key_press = gi.password_widget._widgets[
-        "new_password"
-    ].value_input
-    repeat_new_password_key_press = gi.password_widget._widgets[
-        "repeat_new_password"
-    ].value_input
-    # Check if new password match repeat password
-    if username:
-        if new_password_key_press == repeat_new_password_key_press:
-            # Check if new password is valid with regex
-            if re.fullmatch(
-                config.basic_auth.psw_regex,
-                new_password_key_press,
-            ):
-                # If is_guest and is_admin are None (not passed) use the ones
-                # already set for the user
-                if is_guest is None:
-                    is_guest = auth.is_guest(user=user, config=config)
-                if is_admin is None:
-                    is_admin = auth.is_admin(user=user, config=config)
-                # Add a privileged users only if guest option is not active
-                if not is_guest:
-                    auth.add_privileged_user(
-                        user=username,
-                        is_admin=is_admin,
-                        config=config,
-                    )
-                # Green light: update the password!
-                auth.add_user_hashed_password(
-                    user=username,
-                    password=new_password_key_press,
-                    config=config,
-                )
-
-                # Logout if requested
-                if logout_on_success:
-                    pn.state.notifications.success(
-                        "Password updated<br>Logging out",
-                        duration=config.panel.notifications.duration,
-                    )
-                    sleep(4)
-                    auth.force_logout()
-                else:
-                    pn.state.notifications.success(
-                        "Password updated",
-                        duration=config.panel.notifications.duration,
-                    )
-                return True
-
-            else:
-                pn.state.notifications.error(
-                    "Password requirements not satisfied<br>Check again!",
-                    duration=config.panel.notifications.duration,
-                )
-
-        else:
-            pn.state.notifications.error(
-                "Password are different!",
-                duration=config.panel.notifications.duration,
-            )
-    else:
-        pn.state.notifications.error(
-            "Missing user!",
-            duration=config.panel.notifications.duration,
-        )
-
-    return False
